+++ src/util/cmd_interface.c
@@ -136,18 +136,18 @@ int msg_arg(char *pipe, size_t pipe_len, int cmd, char *arg, size_t arg_len)
   return ret;
 }
 
-int msg_two_arg(char *pipe, int pipe_len, int cmd, char *arg, int arg_len, char *arg2, int arg2_len)
+int msg_two_arg(char *pipe, size_t pipe_len, int cmd, char *arg, size_t arg_len, char *arg2, size_t arg2_len)
 {
   int ret;
   char *msg = (char *) calloc((11 + pipe_len + arg_len + arg2_len), sizeof(char));
 
-  sprintf(msg, "pelz %d %.*s %.*s %.*s", cmd, pipe_len, pipe, arg_len, arg, arg2_len, arg2);
+  sprintf(msg, "pelz %d %.*s %.*s %.*s", cmd, (int)pipe_len, pipe, (int)arg_len, arg, (int)arg2_len, arg2);
   ret = msg_cmd(pipe, msg);
   free(msg);
   return ret;
 }
 
-int msg_list(char *pipe, int pipe_len, int cmd)
+int msg_list(char *pipe, size_t pipe_len, int cmd)
 {
   int ret;
